logica del servicio completada

+++ src/main/java/com/proyectofinal/inventario/repository/RepositorioUsuario.java
@@ -3,7 +3,7 @@ package com.proyectofinal.inventario.repository;
 import com.proyectofinal.inventario.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RepositorioUsuario extends JpaRepository<Usuario,Integer> {
+public interface RepositorioUsuario extends JpaRepository<Usuario,Long> {
 
     Usuario findUsuarioByUserName(String username);
